feat: properly handle/wrap lines in `show` command output
Now, I try to get the terminal window width to know how to show and wrap
long lines. For now, the max width is set to 100 columns, but I'd like
to make it configurable later. 

PS: Note that the wrapping function does not handle ANSI escape codes at
the moment.